break build.py script into two pieces
